fix a few random compiler warnings
 * unused variables
 * implicit enum conversion
 * sigend <-> unsigned comparison
 * left shift negative number

The only warnings left on my setup right now are `-Wstringop-truncation`
from strncpy calls. Those are not trivial to fix. Turns out the code does
regularly not take into account that strncpy may not result in a null
terminated string. Details, see `man string_copying`.

+++ camlibs/ptp2/ptp.c
@@ -412,7 +412,7 @@ fd_putfunc(PTPParams* params, void* private,
 	PTPFDHandlerPrivate* priv = (PTPFDHandlerPrivate*)private;
 
 	written = write (priv->fd, data, sendlen);
-	if (written != sendlen)
+	if ((unsigned long)written != sendlen)
 		return PTP_ERROR_IO;
 	return PTP_RC_OK;
 }
